% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_discontinuous_data.R
\name{generate_discontinuous_data}
\alias{generate_discontinuous_data}
\title{Simulate nonsmooth data}
\usage{
generate_discontinuous_data(n, n_i, k, s_y = 1, s_x = 1, delta_s,
  nt = 101)
}
\arguments{
\item{n}{sample size}

\item{n_i}{number of observations per individual}

\item{k}{tuning parameter which controls the distribution of the parameters that determine the surrogate value}

\item{s_y}{error standard deviation for the outcome}

\item{s_x}{error standard deviation for the surrogate}

\item{delta_s}{residual treatment effect, the parameter of interest}

\item{nt}{number of possible observation times for the longitudinal surrogate}
}
\description{
Simulate nonsmooth data
}
