% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fpc_scores.R
\name{get_fpc_scores}
\alias{get_fpc_scores}
\title{Get FPC scores}
\usage{
get_fpc_scores(ds, ycol, tcol, idcol, options = list(plot = TRUE),
  return_eigenfunction = FALSE)
}
\arguments{
\item{ds}{data.frame containing the data}

\item{ycol}{string denoting the column in ds that contains the longitudinal marker for which FPC scores are desired}

\item{tcol}{string denoting the column in ds that contains the time}

\item{idcol}{string denoting the column in ds that contains the identifier of an individual}

\item{options}{list of options to pass to the FPC fitting procedure}

\item{return_eigenfunction}{logical to indicate whether the eigenfunctions should be returned as well}
}
\description{
Get FPC scores
}
\seealso{
fdapace::FPCA
}
