% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presmooth_data.R
\name{presmooth_data}
\alias{presmooth_data}
\title{Pre-smooth sparse longitudinal data}
\usage{
presmooth_data(obs_data, ...)
}
\arguments{
\item{obs_data}{data.frame or tibble containing the observed data, with columns \code{id} identifying the individual measured, \code{tt} identifying the time of the observation, \code{x} the value of the surrogate at time \code{tt}, and \code{a} indicating 1 for treatment arm and 0 for control arm.}

\item{...}{additional arguments to pass to the function PCA function}
}
\value{
list containing matrices \code{X_t} and \code{X_c}, which are the smoothed surrogate values for the treated and control groups, respectively, for use in downstream analyses
}
\description{
Pre-smooth sparse longitudinal data
}
\examples{
library(dplyr)
library(longsurr)
obs_data <- 
generate_discontinuous_data(n = 50, n_i = 5, delta_s = 0.5, 
k = 1, s_y = 0.1, s_x = 0.1)$obs_ds

head(obs_data)
presmooth_X <- presmooth_data(obs_data)

wide_ds <- full_data \%>\% 
dplyr::select(id, a, tt, x, y) \%>\%
tidyr::spread(tt, x) 

y_t <- wide_ds \%>\%
filter(a == 1) \%>\%
pull(y)
y_c <- wide_ds \%>\%
filter(a == 0) \%>\%
pull(y)
X_t <- presmooth_X$X_t
X_c <- presmooth_X$X_c

estimate_surrogate_value(y_t = y_t, y_c = y_c, 
X_t = X_t, X_c = X_c, method = 'linear')

}
